fix waveform not matching audio file
I added mp3 with generated tones so that I could tell which waveforms belonged to which audio file easily. I realized that the first waveform I clicked on was correct but that all the other waveforms were one behind. I realized that I needed to wait until the metadata was finished loading after changing the audio source before calling the `initPeaks` function.